+++ examples/opencensus/main.go
@@ -1,6 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+// Opencensus exemplifies the use of the OpenCensus to OpenTelemetry bridge.
 package main
 
 import (